requirements.txt
# pip install -r requirements.txt

django 
pillow
django-crispy-forms
django-rest-framework
cx_oracle